feat: refactor GeofenceSelectorModal, CreateWorkLocation, and UpdateWorkLocation for improved unsaved changes handling and remove unused hook
